resume-lava-jobs: Add lava-job cleanup script
This script is needed to inform LAVA server to end the lava-job
process and shutdowns the board/device gracefully while removing
the lava-overlay tmpfs resides on LAVA dispatcher end.

Once lava-job completely endsm it returns a signal to host machine
to continue with the remaining steps in autobuilder before it
completes and end the entire job workflow.

Signed-off-by: Aaron Chan <aaron.chun.yew.chan@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>