config-intelqa-x86_64-lava.json: Add extension to config.json to support BSP
config-intelqa on x86-64 is an extension to config.json where it
contains the recipes and meta layers to build core-image-sato-sdk on
various supported architectures in Yocto Project.

This is an initial release model on specifying the recipes use to build
the embeddded linux images starting with x86-64 MTURBOT64 (Intel IA).
With this reference, community will inherit the structure and model
benchmark from config-intelqa-<arch>-lava.json to support and build
own hardware on other architectures (e.g arm64, mips64, pcc, x86) on
the same common CI infrastructure (Yocto Autobuilder).

The config-intelqa-<arch>-lava.json will contain the consolidate data in
autobuilder and LAVA (Linaro) to execute independent jobs configuration
respectively. Architecture owners are to work with the respective
maintainers and review their automated hardware tests to ensure common
structure is agree by the current and/or new community.

Signed-off-by: Aaron Chan <aaron.chun.yew.chan@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>