config.json: Added meta-openembedded and meta-virtualization
The meta-openembedded has no top-level conf/layer.conf, so
layer-config's attempt to add it fails (silently).  Adding
it to the config.json with no-layer-add set to true will
bypass this silent failure.

The meta-virtualization fails because its repo isn't defined.
Once defined, it has a conf/layer.conf, so it's presence in
NEEDREPOS will cause ADDLAYER to fail because the former will
add it to the bblayers.conf before its dependencies are added
during run-config (causing bitbake-layers to error out).  The
fix is to also specify no-layer-add.

Signed-off-by: Thomas Goodwin <btgoodwin@geontech.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>