config.json: Set BB_NUMBER_PARSE_THREADS
Currently we use 56 or 24 cores on the various workers but this is
pulling machines over with OOM issues, particularly in oe-selftest.
Lower the parallelism for parsing to try and help avoid this as we
don't really need maximum speed for parsing on the autobuilder.

Keep the same numbers for the perf workers as we want to use all
cores there.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>