config.json: Update non-rpm targets to improve coverage
The "non-rpm" tests were not covering our builds particularly well and wasting
build resources. Update the tests to drop the "deb on rpm" and "rpm on deb" targets
which never really find issues.

Replace them with one set of good x86 and arm tests for both ipk and deb, adding in
SDK testing to the mix as well.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>